+++ examples/provider/provider.tf
@@ -2,7 +2,7 @@ terraform {
   required_version = ">= 0.13"
   required_providers {
     b2 = {
-      source  = "Backblaze/b2"
+      source = "Backblaze/b2"
     }
   }
 }