+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Support isFileLockEnabled for buckets
+* Support defaultRetention for buckets
+
+### Fixed
+* Fix acceptance tests breaking when new response fields are added to the API 
+
+### Changed
+* Upgraded b2sdk version to 1.8.0
+
 ## [0.4.0] - 2021-04-08
 
 ### Added